Make pies resizeable and remove titles
