+++ src/routes/portfolio_view.svelte
@@ -8,7 +8,7 @@
 	import equityEmissionsPieData from '../json/data_emissions_pie_equity.json';
 	import techmixData from '../json/data_techexposure.json';
 	import mapData from '../json/data_map.json';
-	import { PieExploded } from '../js/PieExploded2.js';
+	import { PieExploded } from '../js/pie_exploded.js';
 	import { techexposure } from '../js/techexposure';
 	import { choropleth } from '../js/map.js';
